requirements.txt
discord.py
anthropic
openai
python-dotenv
